제출 마감 54분 전
데베 연결 못함

+++ CLIENT/Player.h
@@ -6,8 +6,8 @@ class Player
 
 public:
 	int				id;
-	int				hp;
-	int				max_hp;
+	short			hp;
+	short			max_hp;
 	int				exp;
 	int				level;
 	short			x, y;
